requirements.txt
fastapi==0.119.1
uvicorn==0.37.0
redis==6.4.0
requests==2.32.5
python-dotenv==1.2.1

